A restart of the customizer mod.  All this will do for now is replace the recipe for "Iron Pot" with "Iron Dye Pot" which is a single slot container.
If a dye flask is put inside, using the dye radial menu will dye the invoker's hair the color of the dye.

I will likely expand this to all the other stuff in the character layout that is tintable.

There's another bit of C# code I'll check in shortly that is needed.  There's no functionality to create a new data table at runtime, or write directly to one.  So I export the dye colors as a csv, then the C# program adds 1000 to all of the dye IDs.  I reimport that as a modifier table which I then apply to the hair colors table.  This creates a set of new colors in the hair color table at 1000 and up.

When the dye happens, I look up the dye's table ID, add 1000 and set the hair color ID to that.