Initial checkin of the Booze Helper mod.
I should have had this checked in earlier, as I've done a couple of patches.

Initially, HangOverFixComp would attach via mod controller to the players, and from begin play replace the buff system with a subclass Buffy.  This worked pretty well but any buff that accessed the bp_ac_buffsystem variable from baseCombat would have problems with trying to do things client side.  The server would detect the client trying to do illegal stuff and kick the player off the server.  Really those buffs shouldn't have been trying to do that stuff on the client.

I experimented with alot of variations of having buffy replicate, server only, client and server etc, but the answer suggested by Testerie was to have the mod controller add the component.  Doing that and nuking the old buffsystem and pointing various variables at buffy instead of buffsystem seems to have done the trick.

Oh and putting some of the creation / beginPlay stuff on a delay helped work around a known bug in this version of unreal related to all of this.

So HangOverFixComp will probably go away eventually.

The initial drink taken by a lady would almost always drop the packet sent to fix the sound on the client.  I had to just make that message guaranteed.  I think it can still happen in certain situations.