A user configurable bark byproduct!  Saving the variable was easy, getting the bench to make bark was pretty easy too, way easier than it used to be.
Getting the UI to appear took hours and hours and hours of waiting for the devkit to restart after crashing.  In the end all the assgoblinry was forgetting to set the right parent when I made the widget.

Then getting the value from the UI to set the variable on the server was another like 4 hours.  No editor crashes this time just a blank stare.  All the code paths of all the various attempts I made just to set an integer are still in there for your amusement, but eventually I'll nuke them all and keep the one that works.  They ALL should work, but most act like they work but do nothing.  My favourite.  At least give me an error or something.

In the end I had to make a component on basePlayerChar that I don't even really need just to set a single integer.