A different approach on the camera distance mod.  Simply force set the camera arm length (or target length for the lerp).
It's heavy handed and probably a little expensive but it works.  Left the max distance uncapped but capped minimum as the camera can go negative in vanity mode.

No idea what this will do to archery / horses etc, but it has been up a couple days and no comments so guess it is working?

Need to think about gamepads, this directly reads the mousewheel.